refactor: refactor styles and loading components
