+++ packages/card/proptypes.js
@@ -1,5 +1,15 @@
 import PropTypes from "prop-types";
 
+export const loadingPropTypes = {
+  aspectRatio: PropTypes.number,
+  showImage: PropTypes.bool
+};
+
+export const loadingDefaultProps = {
+  aspectRatio: 3 / 2,
+  showImage: false
+};
+
 export const propTypes = {
   image: PropTypes.shape({ uri: PropTypes.string }),
   imageRatio: PropTypes.number,
