+++ packages/card/styles/shared.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  imageContainer: {
+    marginBottom: 10
+  },
+  headerContainer: {
+    height: 24,
+    marginBottom: 10,
+    maxWidth: 300
+  },
+  textContainer: {
+    height: 10,
+    marginBottom: 10
+  },
+  noMarginBottom: {
+    marginBottom: 0
+  }
+});
